Entry front

+++ frontend/src/components/Dashboard/Profile.jsx
@@ -5,6 +5,7 @@ import ChangePasswordForm from "./Password Modal/PasswordModal";
 import EditProfileModal from "./Edit Profile Modal";
 import EditProfileForm from "./Edit Profile Modal/EditProfileModal";
 import { useNavigate } from "react-router-dom";
+import { toast } from 'react-toastify';
 
 function Profile() {
   const { auth, handleLogout } = useAuth();
@@ -29,28 +30,35 @@ function Profile() {
   };
 
   const handleDeleteProfile = async () => {
-    try {
-      const response = await fetch(`/api/users/${auth.user.id}`, {
-        method: "DELETE",
-        headers: {
-          Authorization: `Bearer ${auth.token}`,
-        },
-      });
 
-      if (response.ok) {
-        alert("Profile deleted successfully");
-        await handleLogout();
-        navigate("/");
-      } else {
-        const errorData = await response.json();
-        const errorMessage = errorData.message || "Profile deletion failed.";
-        alert(errorMessage);
+    const confirmDelete = window.confirm(
+      "Are you sure you want to delete this profile?"
+    );
+
+      if (confirmDelete) {
+      try {
+        const response = await fetch(`/api/users/${auth.user.id}`, {
+          method: "DELETE",
+          headers: {
+            Authorization: `Bearer ${auth.token}`,
+          },
+        });
+
+        if (response.ok) {
+          toast.success("Profile deleted!");
+          await handleLogout();
+          navigate("/");
+        } else {
+          const errorData = await response.json();
+          const errorMessage = errorData.message || "Profile deletion failed.";
+          alert(errorMessage);
+        }
+      } catch (error) {
+        console.error(error);
+        alert("An error occurred during profile deletion.");
       }
-    } catch (error) {
-      console.error(error);
-      alert("An error occurred during profile deletion.");
-    }
-  };
+    };
+  }
 
   console.log(auth.user.email);
   console.log(auth.user.id);
